tpm: split the tpm buffer into separate include file
This moves the tpm buffer implementation into a dedicated header to
enable reuse by the compressed kernel.
